Acoustic/requirements.txt
# Backend
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Audio Processing
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1
numpy==1.26.4
scipy==1.11.4
audioread==3.0.1
noisereduce==2.0.1
torchaudio==2.8.0+cpu
torch==2.8.0+cpu

# Audio Separation / AI Models
spleeter
demucs==4.0.1
openunmix==1.3.0
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
whisper==1.1.10
openai-whisper==20231117

# Utilities
ffmpeg-python
pandas==2.3.1
